+++ library/src/commonMain/kotlin/com/tidal/networktime/internal/NtpUdpSocketOperations.kt
@@ -3,7 +3,7 @@ package com.tidal.networktime.internal
 internal expect class NtpUdpSocketOperations() {
   fun prepareSocket(timeoutMilliseconds: Long)
 
-  fun exchangePacketInPlace(buffer: ByteArray, address: String, portNumber: UInt)
+  fun exchangePacketInPlace(buffer: ByteArray, address: String, portNumber: Byte)
 
   fun closeSocket()
 }
